requirements.txt
facenet_pytorch
pillow
pandas
tqdm